Moved the extra stuff into its own module
